refactor: Moved tests in an archive folder
