+++ tests/test_mails.py
@@ -2,19 +2,26 @@
 # test_mails.py
 # Author: darwh
 # Date: 07/06/2023
-# Description: 
+# Description: Tests for the email system. Manual test but the formatting is done automatically, as well as the
+#              sending of the email, and the content of it is simulated too.
 # -----------------------------------------------------------
 import json
+import os.path
 import unittest
 
-from src.server.data.computer_database import ComputerDatabase
-from src.server.warn_admin.mails import EmailResults
+from src.server.infrastructure.paths import ServerPath
+from src.server.report.mails import EmailResults, setup_email_config, load_email_infos
+from src.server.update_management.network_update_manager import UpdateManager
 
 
 class TestMails(unittest.TestCase):
     def setUp(self) -> None:
-        ComputerDatabase.json_computers_database_filename = "test_computers_database.json"
-        with open(ComputerDatabase.json_computers_database_filename, "w") as file:
+        if not load_email_infos():
+            print("No email config found, creating one to run this test")
+            setup_email_config()
+
+        ServerPath.json_computers_database_filename = "test_computers_database.json"
+        with open(ServerPath.get_database_path(), "w") as file:
             json.dump({
                 "test_computer": {
                     "username": "test_computer\\serveur",
@@ -46,8 +53,14 @@ class TestMails(unittest.TestCase):
                 }
             }, file, indent=4)
 
-        self.database: ComputerDatabase = ComputerDatabase.load_computer_data_if_exists()
+        # BUG: Ça vient de la ligne suivante. Dans load_computer, il obtient le fichier json, mais le mauvais. Il prend
+        # ServerPath et non pas UpdateManager. Voir conv chatGPT.
+        self.database: UpdateManager = UpdateManager.load_computer_data_if_exists()
+        # TODO: Bug ici, computer not found
         computer_mod = self.database.find_computer("test_computer")
+        if computer_mod is None:
+            raise ValueError("Computer not found")
+
         computer_mod.updates_string = [
             '2023-05 Aperçu de la mise à jour cumulative de .NET Framework 3.5',
             '4.8 et 4.8.1 Windows 10 Version 22H2 pour x64(KB5026958)',
@@ -74,3 +87,6 @@ class TestMails(unittest.TestCase):
 
     def test_send_mail(self):
         EmailResults(self.database).send_email_results()
+        database_path = ServerPath.get_database_path()
+        if os.path.exists(ServerPath.get_database_path() and database_path == "test_computers_database.json"):
+            os.remove(ServerPath.get_database_path())
